+++ locallib.php
@@ -1684,8 +1684,7 @@ function theme_boost_union_get_scss_courseoverview_block($theme) {
 
 
 /**
- * Returns the SCSS code to be used in the navbar. The first usage is a logo icon that is possibly too broad and needs
- * reduced to fit the navbar. See theme_boost_union_maxlogowidth for further details
+ * Returns the SCSS code to be used in the navbar.
  *
  * @param theme_config $theme The theme config object.
  * @return string
@@ -1694,9 +1693,12 @@ function theme_boost_union_get_scss_navbar($theme) {
     // Initialize SCSS snippet.
     $scss = '';
 
-    // Set variables which are read in settings by the logo maxwidth values.
-    if (get_config('theme_boost_union', 'maxlogowidth')) {
-        $scss .= ".navbar-brand img.logo{max-width:".get_config('theme_boost_union', 'maxlogowidth').";height:auto;}\n";
+    // Set styles bases on the maxlogowidth setting.
+    if (!empty(get_config('theme_boost_union', 'maxlogowidth'))) {
+        $scss .= '.navbar-brand, .navbar-brand .logo {
+                max-width: '.get_config('theme_boost_union', 'maxlogowidth').';
+                height: auto;
+        }'.PHP_EOL;
     }
 
     return $scss;